+++ src/main/java/com/ganzi/backend/rag/api/RagController.java
@@ -4,8 +4,8 @@ import com.ganzi.backend.global.code.dto.ApiResponse;
 import com.ganzi.backend.global.code.status.SuccessStatus;
 import com.ganzi.backend.global.security.userdetails.CustomUserDetails;
 import com.ganzi.backend.rag.api.doc.RagControllerDoc;
-import com.ganzi.backend.rag.api.dto.RagRequest;
-import com.ganzi.backend.rag.api.dto.RagResponse;
+import com.ganzi.backend.rag.api.dto.request.RagRequest;
+import com.ganzi.backend.rag.api.dto.response.RagResponse;
 import com.ganzi.backend.rag.application.RagService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
