+++ src/main/java/com/ganzi/backend/rag/api/dto/request/RagRequest.java
@@ -1,4 +1,4 @@
-package com.ganzi.backend.rag.api.dto;
+package com.ganzi.backend.rag.api.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
